+++ packages/docs/src/stories/Box.stories.tsx
@@ -1,12 +1,12 @@
 import type { StoryObj, Meta } from '@storybook/react'
-import { Box, BoxProps, Text } from '@allan-koin/react-webview'
+import { Box, BoxProps, Text } from '@allan-koin/react'
 
 export default {
   title: 'Surfaces/Box',
   component: Box,
 
   args: {
-    children: <Text color='$neutral-800'>Testando elemento box</Text>,
+    children: <Text color="$neutral-800">Testando elemento box</Text>,
   },
   argTypes: {
     children: {