+++ packages/react/src/components/ReadOnlyField/index.tsx
@@ -1,0 +1,17 @@
+import { ReadOnlyFieldContainer } from './styles'
+
+interface IProps {
+  title: string
+  text: string
+}
+
+export function ReadOnlyField({ title, text }: IProps) {
+  return (
+    <ReadOnlyFieldContainer>
+      <p>{title}</p>
+      <span>{text}</span>
+    </ReadOnlyFieldContainer>
+  )
+}
+
+ReadOnlyField.displayName = 'ReadOnlyField'