+++ packages/ts-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @allan-koin/ts-config
 
+## 2.3.0
+
+### Minor Changes
+
+- create component readonlyfield
+
 ## 2.2.0
 
 ### Minor Changes